+++ pydidas/data_io/implementations/hdf5_io.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2023, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as
+# published by the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -18,23 +20,23 @@ Module with the Hdf5Io class for importing and exporting Hdf5 data.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2021-2022, Malte Storm, Helmholtz-Zentrum Hereon"
-__license__ = "GPL-3.0"
+__copyright__ = "Copyright 2023, Helmholtz-Zentrum Hereon"
+__license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
-__status__ = "Development"
+__status__ = "Production"
 __all__ = []
 
 import os
 from copy import copy
 from numbers import Integral
 
-from numpy import amax, squeeze
 import h5py
 import hdf5plugin
+from numpy import amax, squeeze
 
-from ...core.constants import HDF5_EXTENSIONS
 from ...core import Dataset
-from ..low_level_readers.read_hdf5_dataset import read_hdf5_dataset
+from ...core.constants import HDF5_EXTENSIONS
+from ..low_level_readers import read_hdf5_dataset
 from .io_base import IoBase